Added: API for Course

+++ handler/courseHandler.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 	"gorm.io/driver/sqlite"
 	"gorm.io/gorm"
+	"ie-backend-project/common"
 	"ie-backend-project/model"
 )
 
@@ -28,23 +29,52 @@ func NewCourseHandler(dsn string) (*CourseHandler, error) {
 
 func (h CourseHandler) NewCourse(name, instructor string) (*model.Course, error) {
 	newCourse := model.Course{Name: name, Instructor: instructor}
+	if h.Exists(newCourse) {
+		return nil, errors.New("duplicate course")
+	}
 	res := h.db.Create(&newCourse)
 	if res.Error != nil {
 		return nil, res.Error
 	}
 	return &newCourse, nil
+}
 
+func (h CourseHandler) AddCourse(course model.Course) (uint, error) {
+	if h.Exists(course) {
+		return 0, common.DuplicateCourseError
+	}
+	res := h.db.Create(&course)
+	if res.Error != nil {
+		return 0, res.Error
+	}
+	return course.ID, nil
 }
 
 func (h CourseHandler) GetCourse(id uint) (*model.Course, error) {
-	course := model.Course{}
-	res := h.db.First(&course, id)
-	if res.Error != nil {
-		return nil, res.Error
+	course := new(model.Course)
+	h.db.Limit(1).Find(course, id)
+	if course.ID != 0 {
+		return course, nil
 	}
-	return &course, nil
+	return nil, common.CourseNotFoundError
+
 }
 
-func (h CourseHandler) DeleteCourse(id uint) {
+func (h CourseHandler) DeleteCourse(id uint) error {
+	course := new(model.Course)
+	h.db.Limit(1).Find(course, id)
+	if course.ID == 0 {
+		return common.CourseNotFoundError
+	}
 	h.db.Delete(&model.Course{}, id)
+	return nil
+}
+
+func (h CourseHandler) Exists(course model.Course) bool {
+	foundOne := new(model.Course)
+	h.db.Where(&model.Course{Name: course.Name, Instructor: course.Instructor}).Limit(1).Find(foundOne)
+	if foundOne.ID != 0 {
+		return true
+	}
+	return false
 }
